cdk_python/requirements.txt
aws-cdk-lib==2.29.1
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha
boto3>=1.24.24
